requirements.txt
sentence-transformers
transformers
pdfminer.six  # often used in custom PDF loaders
numpy
faiss-cpu
fitz
